+++ test/Bridges/lazy_bridge_optimizer.jl
@@ -883,7 +883,7 @@ MOIU.@model(ModelNoZeroIndicator,
              MOI.NormInfinityCone, MOI.NormOneCone,
              MOI.SecondOrderCone, MOI.RotatedSecondOrderCone,
              MOI.GeometricMeanCone, MOI.ExponentialCone, MOI.DualExponentialCone,
-             MOI.NormSpectralCone, MOI.NormNuclearCone,
+             MOI.RelativeEntropyCone, MOI.NormSpectralCone, MOI.NormNuclearCone,
              MOI.PositiveSemidefiniteConeTriangle, MOI.PositiveSemidefiniteConeSquare,
              MOI.RootDetConeTriangle, MOI.RootDetConeSquare, MOI.LogDetConeTriangle,
              MOI.LogDetConeSquare),
@@ -900,7 +900,7 @@ MOIU.@model(ModelNoIndicator,
              MOI.NormInfinityCone, MOI.NormOneCone,
              MOI.SecondOrderCone, MOI.RotatedSecondOrderCone,
              MOI.GeometricMeanCone, MOI.ExponentialCone, MOI.DualExponentialCone,
-             MOI.NormSpectralCone, MOI.NormNuclearCone,
+             MOI.RelativeEntropyCone, MOI.NormSpectralCone, MOI.NormNuclearCone,
              MOI.PositiveSemidefiniteConeTriangle, MOI.PositiveSemidefiniteConeSquare,
              MOI.RootDetConeTriangle, MOI.RootDetConeSquare, MOI.LogDetConeTriangle,
              MOI.LogDetConeSquare),
@@ -1010,6 +1010,8 @@ end
                                       MOI.PositiveSemidefiniteConeSquare)
         @test MOI.supports_constraint(full_bridged_mock, F,
                                       MOI.GeometricMeanCone)
+        @test MOI.supports_constraint(full_bridged_mock, F,
+                                      MOI.RelativeEntropyCone)
         @test !MOI.supports_constraint(
             greater_nonneg_mock, F, MOI.Nonpositives)
         @test MOI.supports_constraint(