+++ NEWS.md
@@ -1,6 +1,11 @@
 MathOptInterface (MOI) release notes
 ====================================
 
+v0.9.11 (February 9, 2020)
+---------------------
+
+- Added `GeoMeantoExpBridge` to bridge a geomean constraint to exponential cone constraints.
+
 v0.9.10 (January 31, 2020)
 ---------------------
 
@@ -14,7 +19,7 @@ v0.9.10 (January 31, 2020)
 - Added `dual_set_type` (#1002).
 - Added tests for vector specialized version of `delete` (#989, #1011).
 - Added PSD3 test (#1007).
-- Clarifed dual solution of `Tests.pow1v` and `Tests.pow1f` (#1013).
+- Clarified dual solution of `Tests.pow1v` and `Tests.pow1f` (#1013).
 - Added support for `EqualTo` and `Zero` in
   `Bridges.Constraint.SplitIntervalBridge` (#1005).
 - Fixed `Utilities.vectorize` for empty vector (#1003).